Issue #27: Saving reconstructed data - don't save large objects
Currently retistruct.save.recdata() saves the r$ol$im field (image)
and the r$ims field (coordinates of corners of pixels in spherical
coordinates), which have the same dimensions as the image, and are
thus large. This patch prevents saving of r$ims. Dealing with r$ol$im
may be dealt with by a future refactoring of image handling.