Issue #48: Check problems on Debian systems
Check problems on the Debian systems are caused by attempts to write
to the user library to which all packages get installed before
checking (and which now is remounted read-only for checking). The
problem was introduced by the testing of retistruct.save.recdata(),
and is rectified by copying the test datasets to tempdir() before
reconstruction.