Fix #49: Error due to change in default stringsAsFactors
The markup data frame was read in using factors and was converted to a
vector with sapply(M.df, function(x) x). However, when strings are
present in M.df, this made every element of the vector a string, which
caused an error later when referring to numeric elements, e.g. phi0.