Ability to use depthmap information
It is now possible to specify the depth at various points of the
flatmount image, for more accurate reconstruction. This is achieved
via a single-channel TIFF file called depthmap.tiff, and a
corresponding CSV file depthmap.csv, which specifies the background
value.

It is also possible to switch easily between 3D views of flatmount and
reconstruction in the GUI.