Stitch outlines comprising multiple fragments with a hole in the centre
The main work has been to generalis the notion of a rim set to a
boundary set. The boundary set is the part of the outline that is not
stitched. It may comprise only one connected part, which is by
definition the rim, or it may comprise multiple closed paths, the
longest of which is assumed to be rim. Once the rim is identified,
reconstruction can start. There may be future work to do with ensuring
the area of the projection is appropriate, but this will probably have
a small influence on the quailty of the reconstruction.