Issue #74: Testing issues with at least one Mac M1 configuration
A floating point issue with Apple clang 17 on the Mac M1 architecture
caused RTriangle to give different triangulations on that architecture
from other architectures. This triangulation caused a reproducibility
test in Retistruct to fail. This issue was fixed in RTriangle 1.6-0.15
(https://github.com/davidcsterratt/RTriangle/issues/23) and we
increase the version of RTriangle required correspondingly.